Create Compose structure
